Upload english version of the guide
I have rewritten my older guide from casualcrypto.date in markdown and
expanded it, as it was not complete. The current text has been corrected
by me and my friend, so it should be mostly correct. This commit
contains a raster image used as a visual description of the line
formation movement proccess.